New functions for unsampled states
- We now have user exposed functions for computing free energies and
expectations in states that have _not_ been sampled, even if they
are not included in the object initialization. The same information
is required (i.e. the recomputed reduced potential energies in the
unsampled states and the observed data for expectations).

- Cleaned up code for reshaping inputs in multiple format - the input
data no longer has to be sorted by state, although this is generally
the easiest way to organize it.

- Expanded some documentation.