Ugly but useful test script
Probably the easiest test case is a collection of harmonic oscillators,
since the free energy and mean coordinates are known analytically. The
hotest.py script contains some pretty ugly tests and profiling code
that can be useful if you really know what you are doing. This is
mostly for myself and ambitious posterity - this should really not be
used or even viewed by casual users.