Numerical computation of Hill equation fits
Experimental(!) estimation of the apparent pKa and Hill coefficient is
now possible directly from the titration curve when using WHAM. This is
because the full curve is known numerically and thus the parameters
can be determined independently of which pH values were sampled (i.e.
there is no recourse to non-linear regression).

For now, this is only implemented for monoprotic, macroscopic systems,
since the equations for polyprotic and microscopic systems are more
complicated and usually involve correlations.