Fix stability bug when estimating pKa
The previous fix for more robust pKa estimation had several flaws
(and was too complicated). It also resulted in totally incorrect
estimates for HIS. The new method is much simpler and should work
in general.

There is also now a catch in cphanalyze to avoid analyzing hopeless
titration curves that are totally flat -- this results in a
substantial performance for blind analysis of a full system.